Allow use existed postgresql server

+++ pytest_postgres/plugin.py
@@ -4,6 +4,7 @@ import uuid
 import docker as docker_client
 import psycopg2
 import pytest
+from docker.errors import APIError, ImageNotFound
 
 
 def pytest_addoption(parser):
@@ -15,6 +16,16 @@ def pytest_addoption(parser):
                      help='Save database container at the end')
     parser.addoption('--pg-network', action='store', default=None,
                      help='Specify docker network for the PostgreSQL container')
+    parser.addoption('--pg-host', action='store', default=None,
+                     help='Specify PostgreSQL server host')
+    parser.addoption('--pg-port', action='store', default=5432,
+                     help='Specify PostgreSQL server port')
+    parser.addoption('--pg-user', action='store', default='postgres',
+                     help='Specify PostgreSQL server user name')
+    parser.addoption('--pg-password', action='store', default='postgres',
+                     help='Specify PostgreSQL server user password')
+    parser.addoption('--pg-database', action='store', default='postgres',
+                     help='Specify test database name')
 
 
 @pytest.fixture(scope='session')
@@ -34,7 +45,7 @@ def check_connection(params):
             time.sleep(delay)
             delay *= 2
     else:
-        pytest.fail('Cannot start postgres server')
+        pytest.fail('Could not connect to PostgreSQL server')
 
 
 def create_container(docker, image, name, ports, network=None):
@@ -49,7 +60,7 @@ def create_container(docker, image, name, ports, network=None):
     if not container:
         try:
             docker.images.pull(image)
-        except docker_client.errors.APIError:
+        except APIError:
             pass
 
         container_params = {'image': image, 'name': name, 'detach': True}
@@ -61,7 +72,7 @@ def create_container(docker, image, name, ports, network=None):
 
         try:
             container = docker.containers.create(**container_params)
-        except docker_client.errors.ImageNotFound:
+        except ImageNotFound:
             pytest.fail('Image `{0}` not found'.format(image))
 
     return container
@@ -69,40 +80,55 @@ def create_container(docker, image, name, ports, network=None):
 
 @pytest.yield_fixture(scope='session')
 def pg_server(docker, request):
+    pg_host = request.config.getoption('--pg-host')
+    pg_port = request.config.getoption('--pg-port')
+    pg_user = request.config.getoption('--pg-user')
+    pg_password = request.config.getoption('--pg-password')
+    pg_database = request.config.getoption('--pg-database')
+
     pg_name = request.config.getoption('--pg-name')
     pg_image = request.config.getoption('--pg-image')
     pg_reuse = request.config.getoption('--pg-reuse')
     pg_network = request.config.getoption('--pg-network')
 
-    if not pg_name:
-        pg_name = 'db-{}'.format(str(uuid.uuid4()))
+    network = None
+    container = None
 
-    container = create_container(docker, pg_image, pg_name, {'5432/tcp': None},
-                                 pg_network)
-    container.start()
-    container.reload()
+    if not pg_host:
+        if not pg_name:
+            pg_name = 'db-{}'.format(str(uuid.uuid4()))
 
-    host = '127.0.0.1'
-    port = 5432
-    if pg_network:
-        net = container.attrs['NetworkSettings']['Networks'][pg_network]
-        host = net['IPAddress']
-    else:
-        ports = container.attrs['NetworkSettings']['Ports']
-        port = ports['5432/tcp'][0]['HostPort']
+        container = create_container(docker, pg_image, pg_name,
+                                     ports={'5432/tcp': None},
+                                     network=pg_network)
+        container.start()
+        container.reload()
+
+        network = container.attrs['NetworkSettings']
+
+        if pg_network:
+            net = container.attrs['NetworkSettings']['Networks'][pg_network]
+            pg_host = net['IPAddress']
+        else:
+            ports = container.attrs['NetworkSettings']['Ports']
+            pg_host = 'localhost'
+            pg_port = ports['5432/tcp'][0]['HostPort']
 
     pg_params = {
-        'database': 'postgres', 'user': 'postgres', 'password': 'postgres',
-        'host': host, 'port': port
+        'host': pg_host,
+        'port': pg_port,
+        'database': pg_database,
+        'user': pg_user,
+        'password': pg_password
     }
 
     try:
         check_connection(pg_params)
         yield {
-            'network': container.attrs['NetworkSettings'],
+            'network': network,
             'params': pg_params
         }
     finally:
-        if not pg_reuse:
+        if not pg_reuse and container:
             container.kill()
             container.remove()
